intmed_test/requirements.txt
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.2
django-allauth==0.38.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-trello-webhooks==0.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
model-mommy==1.6.0
oauthlib==3.0.1
openapi-codec==1.3.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.2
wcwidth==0.1.7